+++ CHANGES.md
@@ -14,6 +14,8 @@
 
 - Added dpi param to plotRaster()
 
+- Added recording of h.t to simData
+
 - Fixed bug: subConn rules now update weight if different weightNorm 
 
 - Fixed bug: missing call to noiseFromRandom() for NetStims